xserver libX11 xcb-proto mesa pixman plink git update 2 Nov 2015
xserver          commit 7d1e4783853f9830344d101ceab087feb19995be
libxcb/xcb-proto commit e500bd837be8366830deb703a7d22d13020fe4c5
libX11           commit dbcb847a08c44d99e4e1de2ba777d63238fb0e03
pixman           commit 3a50806cbeb713b7d1d07a06c6b503d9e43aefe8
mesa             commit af7c98a9c75b17fc8c8ed0989aa732766e5b06d1
putty            commit 323c3c70c98072213b9aff1d04bc91e96632d260

+++ xorg-server/glx/glxcmds.c
@@ -2182,7 +2182,7 @@ __glXDisp_RenderLarge(__GLXclientState * cl, GLbyte * pc)
         __GLXrenderSizeData entry;
         int extra = 0;
         int left = (req->length << 2) - sz_xGLXRenderLargeReq;
-        size_t cmdlen;
+        int cmdlen;
         int err;
 
         /*
